create.c: adapt to new gnulib, use "add_exclude_fp()"
This refactors the "auto-exclude" patch which populates the exclude
structs. This time, we pass the stream file descriptor to exclude.c,
because our directory scanning algorithm no longer does a `chdir()`.

We do a `subfile_open()` which gives us the file descriptor. Keep it
open a little longer and convert to FILE* stream for subsequent call
to `add_exclude_fp`.